+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Changed
+- **Breaking Changed**: `format` and `write` take `Context` trait, this breaks type inference
+- `i` allows access to outer variables
+
 ## [0.4.0] - 2023-03-01
 ### Changed
 - **Breaking Change**: Made error enums `non_exhaustive`